Working homepage slideshow
Added JS and CSS support for slideshow, plus new content slides as samples. Added a link for the interactive shell – Should be replaced with an AJAX action later that loads the Shell into the entire slideshow. Changed the width of the shell iFrame to 100% for better RWD (though and iframe within an iFrame will still present issues if a user resizes the page after the shell loads).

Signed-off-by: J. Hogue <j@projectevolution.com>