Fix vagrant up
Chef was not working, so I wrote a shell script to
provision the box.

Vagrant.sh was designed so if you run it multiple 
times the result will be the same.

Tricks worth mention:

1. ElasticSearch loads on boot.

2. Using pip-accel to save time when package
installation goes wrong.

3. When you `vagrant ssh`, .bashrc will activate
the virtualenv and cd to project’s dir.

4. Some of the fixtures are not working, but it’s
not breaking the script.

5. There is a hack to create a superuser ‘admin’ 
with password ‘admin’.


The goal of the script is to setup everything so 
you can simply `./manage runserver`.