Switch to use non-forked version of django-comments-xtd
The version we've been using doesn't support new migrations
so we can't write a data migration without proper
migration files.

The built-in comments application also doesn't have migrations
so we had to switch to 'django_comments'.

We have to use a forked version because of the following line
in setup.py:

    install_requires=['Django>=1.6']

This would install the latest stable version of Django which
would break our application.