Remove python2 cta (#1219)
* Remove the Python 2 CTA buttons from the downloads page

* Remove 'which version' text from download-banner box fixture

Removes the 'which version' text from the download-banner box fixture
and replaces it with a 'Looking for 2.7? See below' message to direct
visitors to the release list on the downloads page.

This change may have to be manually effected in the CMS in production.

* Remove the Python 2 CTA buttons from the downloads supernav

* Remove Python 2 from homepage downloads box

Removes the 'which version' text from the homepage downloads box as well
as the link for the latest Python 2 version.

* Remove Python 2 from homepage-downloads box fixture

Removes the 'which version' text from the homepage-downloads box fixture
as well as the link to the latest Python 2 version.

This change may have to be manually effected in the CMS in production.