Change parameters order to avoid YUI Compressor bug (#1680)
The previous code lead to an invalid value for the left property:

`left: calc(-50vw+50%);`

This was being caused due to a bug in YUI Compressor. You can read
a detailed explanation about it in this issue:

https://github.com/yui/yuicompressor/issues/59


This is fixed since version 2.4.9-BSI-2. But, until we don't update
prod's YUI Compressor, this PR work as a bugfix since it changes the
parameters order to one properly understood by the bugged version.
More aboute this in this issue comment:

https://github.com/yui/yuicompressor/issues/59#issuecomment-330514588