Sponsor benefit provided assets (#1959)
* add missing migration

* WIP: "Provided Assets" from PSF to sponsors

creates the ability to configure "Provided Assets" associated with a sponsorship that will be fulfilled by the PSF.

The initial "ProvidedTextAsset" is intended to be used for PyCon US 2022 voucher codes, which will be unique to each voucher "type" and sponsorship level.

Additionally, we will likely include "ProvidedFileAsset" that will be used for all sponsorships with Expo Hall benefits to share with them a common "Exhibitor Packet"

* upload files/shared provided file assets for configured benefits

* Make sure delete operations work as expected for Polymorphic models

While testing the PR, I discovered this bug because I couldn't delete
applications I've created. After doing some research, I figured out this
is due to a known bug on django-polymorphic. More on this issue can be
found in this issue:

https://github.com/django-polymorphic/django-polymorphic/issues/229

* Refactor how to reconfigure sponsor benefit to avoid delete operations

* Make sure the assets are also being updated

* Add docstring to make it more clear that the input assets are decoupled from their configuration

* Fix bad context variable name

* Fields that configure relationship between assets and configuration should be read only for editing

Co-authored-by: Bernardo Fontes <bernardoxhc@gmail.com>