Add optional due date to required assets (#1962)
* Make sure delete operations work as expected for Polymorphic models

While testing the PR, I discovered this bug because I couldn't delete
applications I've created. After doing some research, I figured out this
is due to a known bug on django-polymorphic. More on this issue can be
found in this issue:

https://github.com/django-polymorphic/django-polymorphic/issues/229

* Fix identation

* Add new due date column to required assets

* Base command to notify sponsorship applications which have expiring required assets

* Create new notification about required assets close to due date

* Dispatch notifications via management command

* Management command should look for future expiration dates

* Add test to ensure due date within the email context

* Disable asset input if past due date

* Revert "Disable asset input if past due date"

It's OK to allow them to submit after the deadline, as it is mostly for notifying them. If uploads after the deadline are acceptable, the sponsorship team will collect anything put into the boxes

* cast "target_date" to a datetime.date so comparison works

* update styling/wording on forms and emails

* add max_length to RequiredTextAssets, render form to suit

Co-authored-by: Ee Durbin <ewdurbin@gmail.com>