Optimize DB queries to display content object
Django does not allow to select related nested FK from generic fks, such
as content_object. Because of that, a lot of collateral queries were being
executed to display the sponsorship name because it depends on both its package
and sponsor names. This commit changes this by filtering all the sponsorships
and organizing it as a in-memory dict that is only load for this page.