Admin view to list uploaded assets (#1960)
* Bugfix: missing required_asset on querystring to enable post to update single asset

* Bugfix: wrong variable name

* Add new view to list uploaded assets

* Admin list view to list all the assets within the same page

* Enable PSF staff to edit uploaded asset information

* Only tries to fetch for file URL if there's an existing file

* Add custom admin list filter by asset type

* Admin list to filter assets by related sponsorship benefits

* Optimize DB queries to display content object

Django does not allow to select related nested FK from generic fks, such
as content_object. Because of that, a lot of collateral queries were being
executed to display the sponsorship name because it depends on both its package
and sponsor names. This commit changes this by filtering all the sponsorships
and organizing it as a in-memory dict that is only load for this page.

* Admin asset list filter by content type (sponsor or sponsorship)

* Asset list filter by with or without value (uploaded asset)

* Admin should prevent user from adding assets

Since assets are created automatically by the process to handle new sponsorship
applications, having the possibility for users to manually add assets can lead
to states that are unknown to the application. This commit prevents that from
happening.

* Minimal working code to return a zipfile as a HTTP response

* Safety check to avoid exporting partial data

* Refactor to check if asset has value

* More helper properties

* Add text assets to the zipfile

* Use temporary files to add images to zipfile

* Use the sponsor name as the directory name within the zipfile

* Unit test action to export assets as a zipfile

* fix AssetsInline class for admin

* register admin views for File/Response assets

Co-authored-by: Ee Durbin <ewdurbin@gmail.com>