Add Mastodon verification link
The Python Software Foundation has an official Mastodon account, https://fosstodon.org/@ThePSF. The account is not verified. To improve trust with Mastodon users (and the broader fediverse) that the account really is the official account, the python.org website should include a verification link[1]. This PR adds a link rel=“me” tag to the base.html template.

[1] https://joinmastodon.org/verification