Fixup tests that interact with django-allauth ACCOUNT_PREVENT_ENUMERATION
ACCOUNT_PREVENT_ENUMERATION was introduced in django-allauth 0.52.0, and interferes with our expectations.

This should probably be turned on! But for now disable it by default to keep the changeset minimal.

Allauth _used_ to iterate over users to check for email uniquenss but stopped at some point, so we have to create an EmailAdress object for the user in the relevant test-case for duplicate emails