refactor: update `all_day` detection logic
Instead of using a particular resolution on an object, which differs
between `datetime.date` and `datetime.datetime` objects, operate on
parent class of `datetime` and convert all `dt` to timezone-aware
`datetime` values.

This is also in accordance to the model field being a `DateTimeField`,
so we should always be passing the correctly-created object, instead of
a `datetime.date()`, raising `received a naive datetime` warnings.

Removes unused constants.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>